# server/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Alure API
  version: 0.1.0
servers:
  - url: /api/v1
...
paths:
...
  /projects:
    get:
      tags: [projects]
      summary: List projects
      responses:
        '200':
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '200':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{project_id}:
    delete:
      tags: [projects]
      summary: Delete a project (cascade deletes licenses and releases)
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
  /projects/{project_id}/overview:
    get:
      tags: [projects]
      summary: Project overview stats
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Overview stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOverview'
  /projects/{project_id}/modules:
    get:
      tags: [projects]
      summary: List modules for a project
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectModule'
    post:
      tags: [projects]
      summary: Create a module for a project
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectModuleRequest'
      responses:
        '200':
          description: Project module
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectModule'
  /projects/{project_id}/modules/{module_id}:
    put:
      tags: [projects]
      summary: Update a project module
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
        - in: path
          name: module_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectModuleRequest'
      responses:
        '200':
          description: Project module
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectModule'
    delete:
      tags: [projects]
      summary: Delete a project module
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
        - in: path
          name: module_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
  /projects/{project_id}/releases:
    get:
      tags: [projects]
      summary: List releases for a project
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Release list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReleaseListItem'
    post:
      tags: [projects]
      summary: Create a release for a project
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReleaseRequest'
      responses:
        '200':
          description: Release created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseListItem'
  /projects/{project_id}/releases/upload:
    post:
      tags: [projects]
      summary: Upload a release asset and create the release
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, version, channel]
              properties:
                file:
                  type: string
                  format: binary
                version:
                  type: string
                channel:
                  type: string
                  enum: [stable, beta, hotfix]
                notes:
                  type: string
      responses:
        '200':
          description: Release created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseListItem'
  /projects/{project_id}/releases/{release_id}:
    get:
      tags: [projects]
      summary: Get release details
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
        - in: path
          name: release_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Release detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseListItem'
  /projects/{project_id}/releases/{release_id}/promote:
    post:
      tags: [projects]
      summary: Promote or rollback a release to another channel
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
        - in: path
          name: release_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoteReleaseRequest'
      responses:
        '200':
          description: Release promoted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseListItem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    Project:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectOverview:
      type: object
      required: [licenses, releases, activations, modules]
      properties:
        licenses:
          $ref: '#/components/schemas/ProjectOverviewLicenses'
        releases:
          $ref: '#/components/schemas/ProjectOverviewReleases'
        activations:
          $ref: '#/components/schemas/ProjectOverviewActivations'
        modules:
          $ref: '#/components/schemas/ProjectOverviewModules'
    ProjectOverviewLicenses:
      type: object
      required: [total, active, revoked, expired, avg_usage_percent]
      properties:
        total:
          type: integer
        active:
          type: integer
        revoked:
          type: integer
        expired:
          type: integer
        avg_usage_percent:
          type: integer
    ProjectOverviewReleases:
      type: object
      required: [total]
      properties:
        total:
          type: integer
        latest_version:
          type: string
        latest_channel:
          type: string
        latest_published_at:
          type: string
          format: date-time
    ProjectOverviewActivations:
      type: object
      required: [total, active, revoked]
      properties:
        total:
          type: integer
        active:
          type: integer
        revoked:
          type: integer
    ProjectOverviewModules:
      type: object
      required: [total, active_in_licenses, forced_on, forced_off, used_in_activations]
      properties:
        total:
          type: integer
        active_in_licenses:
          type: integer
        forced_on:
          type: integer
        forced_off:
          type: integer
        used_in_activations:
          type: integer
    ProjectModule:
      type: object
      required: [id, key, name, created_at]
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        description:
          type: string
        params:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    CreateProjectModuleRequest:
      type: object
      required: [key, name]
      properties:
        key:
          type: string
        name:
          type: string
        description:
          type: string
        params:
          type: object
          additionalProperties: true
    UpdateProjectModuleRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        params:
          type: object
          additionalProperties: true
    CreateReleaseRequest:
      type: object
      required: [version, channel]
      properties:
        version:
          type: string
        channel:
          type: string
          enum: [stable, beta, hotfix]
        notes:
          type: string
        asset:
          type: object
          required: [filename, size_bytes, sha256]
          properties:
            filename:
              type: string
            size_bytes:
              type: integer
            sha256:
              type: string
    PromoteReleaseRequest:
      type: object
      required: [channel]
      properties:
        channel:
          type: string
          enum: [stable, beta, hotfix]
    ReleaseListItem:
      type: object
      required: [id, version, channel, published_at]
      properties:
        id:
          type: string
        version:
          type: string
        channel:
          type: string
          enum: [stable, beta, hotfix]
        notes:
          type: string
        published_at:
          type: string
          format: date-time
        asset:
          type: object
          required: [id, filename, size_bytes, sha256, download_url]
          properties:
            id:
              type: string
            filename:
              type: string
            size_bytes:
              type: integer
            sha256:
              type: string
            download_url:
              type: string
...